+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - get a specific customer by id
 - get and delete rated items
 - get a specific payment transaction by id
+- call to terminate a contract
 
 ### Changed
 - update version of jruby to 9.1.5.0